stream reader test refactor
