+++ tests/component/stream_readers/__init__.py
@@ -1,0 +1,1 @@
+"""Stream readers test module."""
