+++ tests/component/stream_readers/test_analog_unscaled_reader.py
@@ -1,0 +1,148 @@
+from __future__ import annotations
+
+import ctypes
+import math
+
+import numpy
+import pytest
+
+import nidaqmx
+from nidaqmx.stream_readers import AnalogUnscaledReader
+from ..conftest import (
+    VOLTAGE_CODE_EPSILON,
+    _get_voltage_code_offset_for_chan,
+)
+
+
+def test___analog_unscaled_reader___read_int16___returns_valid_samples(
+    ai_multi_channel_task: nidaqmx.Task,
+) -> None:
+    reader = AnalogUnscaledReader(ai_multi_channel_task.in_stream)
+    num_channels = ai_multi_channel_task.number_of_channels
+    samples_to_read = 10
+    data = numpy.full(
+        (num_channels, samples_to_read), numpy.iinfo(numpy.int16).min, dtype=numpy.int16
+    )
+
+    samples_read = reader.read_int16(data, number_of_samples_per_channel=samples_to_read)
+
+    assert samples_read == samples_to_read
+    expected_vals = [
+        _get_voltage_code_offset_for_chan(chan_index) for chan_index in range(num_channels)
+    ]
+    assert data == pytest.approx(expected_vals, abs=VOLTAGE_CODE_EPSILON)
+
+
+def test___analog_unscaled_reader___read_int16___raises_error_with_correct_dtype(
+    ai_multi_channel_task: nidaqmx.Task,
+) -> None:
+    reader = AnalogUnscaledReader(ai_multi_channel_task.in_stream)
+    num_channels = ai_multi_channel_task.number_of_channels
+    samples_to_read = 10
+    data = numpy.full((num_channels, samples_to_read), math.inf, dtype=numpy.float64)
+
+    with pytest.raises((ctypes.ArgumentError, TypeError)) as exc_info:
+        _ = reader.read_int16(data, number_of_samples_per_channel=samples_to_read)
+
+    assert "int16" in exc_info.value.args[0]
+
+
+def test___analog_unscaled_reader___read_uint16___returns_valid_samples(
+    ai_multi_channel_task: nidaqmx.Task,
+) -> None:
+    reader = AnalogUnscaledReader(ai_multi_channel_task.in_stream)
+    num_channels = ai_multi_channel_task.number_of_channels
+    samples_to_read = 10
+    data = numpy.full(
+        (num_channels, samples_to_read), numpy.iinfo(numpy.uint16).min, dtype=numpy.uint16
+    )
+
+    samples_read = reader.read_uint16(data, number_of_samples_per_channel=samples_to_read)
+
+    assert samples_read == samples_to_read
+    expected_vals = [
+        _get_voltage_code_offset_for_chan(chan_index) for chan_index in range(num_channels)
+    ]
+    # Promote to larger signed type to avoid overflow w/NumPy 2.0+.
+    assert data.astype(numpy.int32) == pytest.approx(expected_vals, abs=VOLTAGE_CODE_EPSILON)
+
+
+def test___analog_unscaled_reader___read_uint16___raises_error_with_correct_dtype(
+    ai_multi_channel_task: nidaqmx.Task,
+) -> None:
+    reader = AnalogUnscaledReader(ai_multi_channel_task.in_stream)
+    num_channels = ai_multi_channel_task.number_of_channels
+    samples_to_read = 10
+    data = numpy.full((num_channels, samples_to_read), math.inf, dtype=numpy.float64)
+
+    with pytest.raises((ctypes.ArgumentError, TypeError)) as exc_info:
+        _ = reader.read_uint16(data, number_of_samples_per_channel=samples_to_read)
+
+    assert "uint16" in exc_info.value.args[0]
+
+
+def test___analog_unscaled_reader___read_int32___returns_valid_samples(
+    ai_multi_channel_task: nidaqmx.Task,
+) -> None:
+    reader = AnalogUnscaledReader(ai_multi_channel_task.in_stream)
+    num_channels = ai_multi_channel_task.number_of_channels
+    samples_to_read = 10
+    data = numpy.full(
+        (num_channels, samples_to_read), numpy.iinfo(numpy.int32).min, dtype=numpy.int32
+    )
+
+    samples_read = reader.read_int32(data, number_of_samples_per_channel=samples_to_read)
+
+    assert samples_read == samples_to_read
+    expected_vals = [
+        _get_voltage_code_offset_for_chan(chan_index) for chan_index in range(num_channels)
+    ]
+    assert data == pytest.approx(expected_vals, abs=VOLTAGE_CODE_EPSILON)
+
+
+def test___analog_unscaled_reader___read_int32___raises_error_with_correct_dtype(
+    ai_multi_channel_task: nidaqmx.Task,
+) -> None:
+    reader = AnalogUnscaledReader(ai_multi_channel_task.in_stream)
+    num_channels = ai_multi_channel_task.number_of_channels
+    samples_to_read = 10
+    data = numpy.full((num_channels, samples_to_read), math.inf, dtype=numpy.float64)
+
+    with pytest.raises((ctypes.ArgumentError, TypeError)) as exc_info:
+        _ = reader.read_int32(data, number_of_samples_per_channel=samples_to_read)
+
+    assert "int32" in exc_info.value.args[0]
+
+
+def test___analog_unscaled_reader___read_uint32___returns_valid_samples(
+    ai_multi_channel_task: nidaqmx.Task,
+) -> None:
+    reader = AnalogUnscaledReader(ai_multi_channel_task.in_stream)
+    num_channels = ai_multi_channel_task.number_of_channels
+    samples_to_read = 10
+    data = numpy.full(
+        (num_channels, samples_to_read), numpy.iinfo(numpy.uint32).min, dtype=numpy.uint32
+    )
+
+    samples_read = reader.read_uint32(data, number_of_samples_per_channel=samples_to_read)
+
+    assert samples_read == samples_to_read
+    expected_vals = [
+        _get_voltage_code_offset_for_chan(chan_index) for chan_index in range(num_channels)
+    ]
+    # Promote to larger signed type to avoid overflow w/NumPy 2.0+.
+    assert data.astype(numpy.int64) == pytest.approx(expected_vals, abs=VOLTAGE_CODE_EPSILON)
+
+
+def test___analog_unscaled_reader___read_uint32___raises_error_with_correct_dtype(
+    ai_multi_channel_task: nidaqmx.Task,
+) -> None:
+    reader = AnalogUnscaledReader(ai_multi_channel_task.in_stream)
+    num_channels = ai_multi_channel_task.number_of_channels
+    samples_to_read = 10
+    data = numpy.full((num_channels, samples_to_read), math.inf, dtype=numpy.float64)
+
+    with pytest.raises((ctypes.ArgumentError, TypeError)) as exc_info:
+        _ = reader.read_uint32(data, number_of_samples_per_channel=samples_to_read)
+
+    assert "uint32" in exc_info.value.args[0]
