+++ tests/component/test_task_read_ai.py
@@ -3,50 +3,15 @@ from __future__ import annotations
 import pytest
 
 import nidaqmx
-import nidaqmx.system
 from nidaqmx.constants import AcquisitionType
-
-
-# Simulated DAQ voltage data is a noisy sinewave within the range of the minimum and maximum values
-# of the virtual channel. We can leverage this behavior to validate we get the correct data from
-# the Python bindings.
-def _get_voltage_offset_for_chan(chan_index: int) -> float:
-    return float(chan_index + 1)
-
-
-VOLTAGE_EPSILON = 1e-3
-
-
-@pytest.fixture
-def ai_single_channel_task(
-    task: nidaqmx.Task, sim_6363_device: nidaqmx.system.Device
-) -> nidaqmx.Task:
-    offset = _get_voltage_offset_for_chan(0)
-    task.ai_channels.add_ai_voltage_chan(
-        sim_6363_device.ai_physical_chans[0].name,
-        min_val=offset,
-        max_val=offset + VOLTAGE_EPSILON,
-    )
-    return task
-
-
-@pytest.fixture
-def ai_multi_channel_task(
-    task: nidaqmx.Task, sim_6363_device: nidaqmx.system.Device
-) -> nidaqmx.Task:
-    for chan_index in range(3):
-        offset = _get_voltage_offset_for_chan(chan_index)
-        chan = task.ai_channels.add_ai_voltage_chan(
-            sim_6363_device.ai_physical_chans[chan_index].name,
-            min_val=offset,
-            # min and max must be different, so add a small epsilon
-            max_val=offset + VOLTAGE_EPSILON,
-        )
-        # forcing the maximum range for binary read scaling to be predictable
-        chan.ai_rng_high = 10
-        chan.ai_rng_low = -10
-
-    return task
+from .conftest import (
+    POWER_EPSILON,
+    VOLTAGE_EPSILON,
+    _assert_equal_2d,
+    _get_current_setpoint_for_chan,
+    _get_voltage_offset_for_chan,
+    _get_voltage_setpoint_for_chan,
+)
 
 
 def test___analog_single_channel___read_unset_samples___returns_valid_scalar(
@@ -151,51 +116,6 @@ def test___analog_multi_channel_finite___read_too_many_sample___returns_valid_2d
         for chan_index in range(num_channels)
     ]
     _assert_equal_2d(data, expected, abs=VOLTAGE_EPSILON)
-
-
-def _assert_equal_2d(data: list[list[float]], expected: list[list[float]], abs: float) -> None:
-    # pytest.approx() does not support nested data structures.
-    assert len(data) == len(expected)
-    for i in range(len(data)):
-        assert data[i] == pytest.approx(expected[i], abs=abs)
-
-
-POWER_EPSILON = 1e-3
-
-
-def _get_voltage_setpoint_for_chan(chan_index: int) -> float:
-    return float(chan_index + 1)
-
-
-def _get_current_setpoint_for_chan(chan_index: int) -> float:
-    return float(chan_index + 1)
-
-
-@pytest.fixture
-def pwr_single_channel_task(
-    task: nidaqmx.Task, sim_ts_power_device: nidaqmx.system.Device
-) -> nidaqmx.Task:
-    task.ai_channels.add_ai_power_chan(
-        f"{sim_ts_power_device.name}/power",
-        _get_voltage_setpoint_for_chan(0),
-        _get_current_setpoint_for_chan(0),
-        True,  # output enable
-    )
-    return task
-
-
-@pytest.fixture
-def pwr_multi_channel_task(
-    task: nidaqmx.Task, sim_ts_power_devices: list[nidaqmx.system.Device]
-) -> nidaqmx.Task:
-    for chan_index, sim_ts_power_device in enumerate(sim_ts_power_devices):
-        task.ai_channels.add_ai_power_chan(
-            f"{sim_ts_power_device.name}/power",
-            _get_voltage_setpoint_for_chan(chan_index),
-            _get_current_setpoint_for_chan(chan_index),
-            True,  # output enable
-        )
-    return task
 
 
 def test___power_single_channel___read_unset_samples___returns_valid_scalar(
